Q11: Print bool as int

+++ questions/011-function-pointer-comparison.rs
@@ -4,5 +4,5 @@ fn g<'a: 'a>() {}
 fn main() {
     let pf = f::<'static> as fn();
     let pg = g::<'static> as fn();
-    print!("{}", pf == pg);
+    print!("{}", (pf == pg) as u8);
 }
